move collection enum to constants/collection
