+++ src/constants/collection/index.ts
@@ -1,3 +1,15 @@
 export const ARCHIVE_SECTION = -1;
 export const TRASH_SECTION = -2;
 export const ALL_SECTION = 0;
+
+export enum CollectionType {
+    folder = 'folder',
+    favorites = 'favorites',
+    album = 'album',
+}
+
+export enum COLLECTION_SORT_BY {
+    LATEST_FILE,
+    MODIFICATION_TIME,
+    NAME,
+}
